requirements.txt
absl-py==1.4.0
anyio==3.7.0
astunparse==1.6.3
boto3==1.26.142
botocore==1.29.142
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cognitojwt==1.4.1
cryptography==40.0.2
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-cognito==2.3.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
idna==3.4
jax==0.4.10
jmespath==1.0.1
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
mkl-random
mkl-service==2.4.0
ml-dtypes==0.1.0
numpy
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.4
wrapt==1.14.1
